requirements.txt
selenium==4.11.2
pytest==7.3.1
allure-pytest==2.13.1